Don't convey positional parameters to the . builtin
To do so is a bashism; POSIX makes no provision for it. Fortunately, there
is no need to begin with. The srcdir variable is already defined at the
point that build/version.sh is sourced from the configure script and, thus,
does not need to be passed.

As version.sh is only ever sourced, that would be a valid fix in and as of
itself. However, because version.sh has the executable bit set, there is an
implication that the original author may have intended for it to be so; for
testing purposes, perhaps. So as to continue to support this potential use
case, have the script retain the ability to assign the value of $1 to
srcdir, provided that srcdir is empty and provided that at least one
positional parameter was specified.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Closes: https://bugs.gentoo.org/764881