Disable timestamp shifting in subtitle-only MKV files
For normal media file, the starting timestamp is shifted to 0. This is necessary in order to playback trimmed audio/video that has a non-zero starting timestamp.
However, if the file only contains subtitle, we do not want to shift the start time of first line to 0. Because subtitles are never played alone, but usually attached to an audio/video. There is no way to get the starting timestamp of the audio/video in a subtitle-only MKV, so assuming 0 is a safe guess.

Fix TypesettingTools/Aegisub#91